+++ src/commands/season.ts
@@ -4,7 +4,7 @@ import { RichEmbed } from "discord.js";
 import { ICommand } from "../interfaces/ICommand";
 import { consts } from "../localization";
 
-import { boldIF } from "../helpers";
+import { boldIF } from "../helpers/functions";
 
 module.exports = {
   name: "seasoninfo",
